@@ -25,8 +25,8 @@ mdx_truly_sane_lists = { version = "1.3", optional = true }
 # Same as Black.
 line-length = 100
 
-# Assume Python 3.9
-target-version = "py39"
+# Assume Python 3.10
+target-version = "py310"
 
 # In addition to the standard set of exclusions, omit all tests, plus a specific file.
 extend-exclude = ["*/.*/",".*/","__pycache__","node_modules"]
